Minor refactoring of shared functions.
