+++ cmd/tuf/app/sign.go
@@ -21,14 +21,11 @@ package app
 import (
 	"bytes"
 	"context"
-	"crypto"
 	"flag"
 	"fmt"
 	"strings"
 
 	"github.com/peterbourgon/ff/v3/ffcli"
-	"github.com/sigstore/cosign/pkg/cosign/pivkey"
-	csignature "github.com/sigstore/cosign/pkg/signature"
 	"github.com/sigstore/root-signing/pkg/keys"
 	"github.com/sigstore/root-signing/pkg/repo"
 	"github.com/sigstore/sigstore/pkg/signature"
@@ -82,7 +79,7 @@ func Sign() *ffcli.Command {
 			if !*sk && *key == "" {
 				return flag.ErrHelp
 			}
-			signer, err := getSigner(ctx, *sk, *key)
+			signer, err := GetSigner(ctx, *sk, *key)
 			if err != nil {
 				return err
 			}
@@ -125,32 +122,6 @@ func checkMetaForRole(store tuf.LocalStore, role []string) error {
 		}
 	}
 	return nil
-}
-
-func getSigner(ctx context.Context, sk bool, keyRef string) (signature.Signer, error) {
-	if sk {
-		pivKey, err := pivkey.GetKeyWithSlot("signature")
-		if err != nil {
-			return nil, err
-		}
-		return pivKey.SignerVerifier()
-	}
-	// A key reference was provided.
-	// First try to load it as a regular PEM encoded private key.
-	signer, err := signature.LoadSignerFromPEMFile(keyRef, crypto.SHA256, nil)
-	if err != nil {
-		var innerError error
-		signer, innerError = csignature.SignerVerifierFromKeyRef(ctx, keyRef, nil)
-		if innerError != nil {
-			// Only print this message if both attempts failed.
-			// As there is a natual fallthrough here, always
-			// logging the first error could be noisy.
-			fmt.Printf("failed to load key as PEM encoded: %s, trying other methods: ", err)
-			return nil, innerError
-		}
-
-	}
-	return signer, nil
 }
 
 func SignCmd(ctx context.Context, directory string, roles []string, signer signature.Signer,
